+++ src/cmd/browse/templates/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod layout;
 mod page_template;
 
+pub use layout::*;
 pub use page_template::*;
