+++ src/cmd/browse/views/card_cloze.rs
@@ -21,8 +21,6 @@ use maud::PreEscaped;
 use maud::html;
 
 use crate::cmd::browse::entries::EntryKey;
-use crate::cmd::browse::layout::Selection;
-use crate::cmd::browse::layout::columns_page;
 use crate::cmd::browse::render::performance_rows;
 use crate::cmd::browse::render::render_config;
 use crate::cmd::browse::render::render_family_revealed;
@@ -32,6 +30,8 @@ use crate::cmd::browse::shared::BrowseState;
 use crate::cmd::browse::shared::error_response;
 use crate::cmd::browse::shared::internal_error_response;
 use crate::cmd::browse::shared::ok_response;
+use crate::cmd::browse::templates::Selection;
+use crate::cmd::browse::templates::columns_page;
 use crate::error::ErrorReport;
 use crate::error::Fallible;
 use crate::error::fail;
