+++ src/cmd/browse/views/deck.rs
@@ -17,12 +17,12 @@ use axum::extract::State;
 use axum::http::StatusCode;
 use axum::response::Html;
 
-use crate::cmd::browse::layout::Selection;
-use crate::cmd::browse::layout::columns_page;
 use crate::cmd::browse::shared::BrowseState;
 use crate::cmd::browse::shared::error_response;
 use crate::cmd::browse::shared::internal_error_response;
 use crate::cmd::browse::shared::ok_response;
+use crate::cmd::browse::templates::Selection;
+use crate::cmd::browse::templates::columns_page;
 
 pub async fn deck_handler(
     State(state): State<BrowseState>,
