+++ src/cmd/browse/views/index.rs
@@ -16,11 +16,11 @@ use axum::extract::State;
 use axum::http::StatusCode;
 use axum::response::Html;
 
-use crate::cmd::browse::layout::Selection;
-use crate::cmd::browse::layout::columns_page;
 use crate::cmd::browse::shared::BrowseState;
 use crate::cmd::browse::shared::internal_error_response;
 use crate::cmd::browse::shared::ok_response;
+use crate::cmd::browse::templates::Selection;
+use crate::cmd::browse::templates::columns_page;
 
 pub async fn index_handler(State(state): State<BrowseState>) -> (StatusCode, Html<String>) {
     let selection = Selection {
